modified models to use dynamodb-enhanced dependency
