--- src/main/java/com/revature/post_user_favorites/SetDocument.java
@@ -1,21 +1,0 @@
-package com.revature.post_user_favorites;
-
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBDocument;
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@Builder
-@DynamoDBDocument
-public class SetDocument {
-    private String id;
-    private String name;
-    private List<Tag> tags;
-    private boolean isPublic;
-    private int views;
-    private int plays;
-    private int studies;
-    private int favorites;
-}
