--- src/main/java/com/revature/post_user_favorites/Tag.java
@@ -1,11 +1,0 @@
-package com.revature.post_user_favorites;
-
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBDocument;
-import lombok.Data;
-
-@Data
-@DynamoDBDocument
-public class Tag {
-    private String tagName;
-    private String tagColor;
-}
